pyproject.toml
[project]
name = "piepdf"
version = "0.1.0"
description = "Manage Research Articles and fetch metadata from the internet"

readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "habanero>=2.0.0",
    "pyqt5>=5.15.11",
    "pyqt5-sip>=12.16.0",
    "pyqtwebengine>=5.15.7",
    "python-poppler-qt5>=21.3.0",
    "sip>=6.9.0",
    "feedparser>=6.0.11",
    "httplib2>=0.22.0",
    "pyqt5-qt5>=5.15.18",
    "pyqt6>=6.10.2",
]

[project.scripts]
piepdf = "piepdf.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61","wheel"]
build-backend = "setuptools.build_meta"
